refactor(web): decompose finyk Assets.tsx into smaller modules
Split finyk/pages/Assets.tsx (872 LOC) into:
- useAssetsState.ts (259 LOC) — custom hook: state, computed values, effects
- AssetsForm.tsx (376 LOC) — SubscriptionForm, AssetForm, ReceivableForm, DebtForm
- AssetsTable.tsx (511 LOC) — AssetsNetworthCard, section components, main table
- Assets.tsx (40 LOC) — thin orchestrator with re-exports

All files under 600 LOC. Zero behaviour changes — pure decomposition.

Added unit tests:
- useAssetsState.test.ts — 9 tests (computed values, section toggle, quick-actions)
- AssetsTable.test.tsx — 4 tests (networth card render/hide/bar)
- AssetsForm.test.tsx — 6 tests (form render, cancel handlers)

Updated docs/frontend-tech-debt.md: removed Assets.tsx from top-LOC list (24 files remain).

Ref: docs/audits/2026-04-26-sergeant-audit-devin.md block 3, PR-3.B
