+++ apps/web/src/modules/finyk/pages/Assets.tsx
@@ -1,225 +1,31 @@
-import { useEffect, useMemo, useRef, useState } from "react";
-import { DebtCard } from "../components/DebtCard";
-import { SubCard } from "../components/SubCard";
-import { RecurringSuggestions } from "../components/RecurringSuggestions";
-import { SectionHeading } from "@shared/components/ui/SectionHeading";
-import { Button } from "@shared/components/ui/Button";
-import { Card } from "@shared/components/ui/Card";
-import { Input } from "@shared/components/ui/Input";
-import { Icon } from "@shared/components/ui/Icon";
-import {
-  getAccountLabel,
-  getMonoDebt,
-  isMonoDebt,
-  getDebtPaid,
-  getRecvPaid,
-  getMonoTotals,
-  calcDebtRemaining,
-  calcReceivableRemaining,
-  getDebtEffectiveTotal,
-  getReceivableEffectiveTotal,
-} from "../utils";
-import { filterVisibleAccounts } from "@sergeant/finyk-domain/domain/assets/aggregates";
-import { cn } from "@shared/lib/cn";
-import { openHubModule } from "@shared/lib/hubNav";
-import { notifyFinykRoutineCalendarSync } from "../hubRoutineSync";
-import { VoiceMicButton } from "@shared/components/ui/VoiceMicButton";
-import { parseExpenseSpeech as parseExpenseVoice } from "@sergeant/shared";
-import { computeFinykSchedule, startOfToday } from "../lib/upcomingSchedule";
-import { FinykStatsStrip } from "../components/FinykStatsStrip";
-import {
-  AssetsLiabilitiesBar,
-  QuickActionButton,
-  SectionBar,
-} from "./AssetsBars";
+import { useAssetsState, type AssetsProps } from "./useAssetsState";
 import { AssetsTxPickerView } from "./AssetsTxPickerView";
+import { AssetsTable } from "./AssetsTable";
 
 export function Assets({
   mono,
   storage,
   showBalance = true,
   initialOpenDebt = false,
-}) {
-  const { accounts, transactions } = mono;
-  const {
-    hiddenAccounts,
-    manualAssets,
-    setManualAssets,
-    manualDebts,
-    setManualDebts,
-    receivables,
-    setReceivables,
-    toggleLinkedTx,
-    subscriptions,
-    setSubscriptions,
-    updateSubscription,
-    addSubscriptionFromRecurring,
-    dismissedRecurring,
-    dismissRecurring,
-    excludedTxIds,
-    monoDebtLinkedTxIds,
-    toggleMonoDebtTx,
-    customCategories,
-  } = storage;
+}: AssetsProps) {
+  const state = useAssetsState({ mono, storage, showBalance, initialOpenDebt });
 
-  const [showAssetForm, setShowAssetForm] = useState(false);
-  const [showDebtForm, setShowDebtForm] = useState(initialOpenDebt);
-  const [showRecvForm, setShowRecvForm] = useState(false);
-  const [showSubForm, setShowSubForm] = useState(false);
-  const [newAsset, setNewAsset] = useState({
-    name: "",
-    amount: "",
-    currency: "UAH",
-    emoji: "💰",
-  });
-  const [newDebt, setNewDebt] = useState({
-    name: "",
-    emoji: "💸",
-    totalAmount: "",
-    dueDate: "",
-  });
-  const [newRecv, setNewRecv] = useState({
-    name: "",
-    emoji: "👤",
-    amount: "",
-    note: "",
-    dueDate: "",
-  });
-  const [newSub, setNewSub] = useState<{
-    name: string;
-    emoji: string;
-    keyword: string;
-    billingDay: string | number;
-    currency: string;
-  }>({
-    name: "",
-    emoji: "📱",
-    keyword: "",
-    billingDay: "",
-    currency: "UAH",
-  });
-  const [txPicker, setTxPicker] = useState(null);
-  const [open, setOpen] = useState({
-    subscriptions: false,
-    assets: false,
-    liabilities: initialOpenDebt,
-  });
-  const assetFormRef = useRef<HTMLElement | null>(null);
-  const assetNameInputRef = useRef<HTMLInputElement | null>(null);
-  const debtFormRef = useRef<HTMLElement | null>(null);
-  const debtNameInputRef = useRef<HTMLInputElement | null>(null);
-
-  const { balance: monoTotal, debt: monoTotalDebt } = getMonoTotals(
-    accounts,
-    hiddenAccounts,
-  );
-  // Список debt-карток у UI мусить бити узгоджений із `monoTotalDebt`
-  // — інакше схована кредитка не контрибує у борг, але все одно
-  // показується у Liabilities-секції з власним remaining-сумамою.
-  const monoDebtAccounts = filterVisibleAccounts(
-    accounts,
-    hiddenAccounts,
-  ).filter((a) => isMonoDebt(a));
-  const manualDebtTotal = manualDebts.reduce(
-    (s, d) => s + calcDebtRemaining(d, transactions),
-    0,
-  );
-  const totalDebt = monoTotalDebt + manualDebtTotal;
-  const totalReceivable = receivables.reduce(
-    (s, r) => s + calcReceivableRemaining(r, transactions),
-    0,
-  );
-  const manualAssetTotal = manualAssets
-    .filter((a) => a.currency === "UAH")
-    .reduce((s, a) => s + Number(a.amount), 0);
-  const networth = monoTotal + manualAssetTotal + totalReceivable - totalDebt;
-  const totalAssets = monoTotal + manualAssetTotal + totalReceivable;
-
-  // Stats strip + upcoming-charge feed. `todayStart` is pinned to a
-  // mount-time midnight via lazy initialiser so `useMemo` deps stay
-  // referentially stable for the session — date only matters for
-  // day-level comparisons, so a frozen snapshot is safer than a fresh
-  // `new Date()` each render.
-  const [todayStart] = useState<Date>(startOfToday);
-
-  // На Активи з усього schedule потрібен тільки `urgentLiability` —
-  // інші тайли (`subsMonthly` / `nextCharge`) переїхали на Планування,
-  // щоб не дублювати ту саму інформацію на двох сторінках.
-  const { urgentLiability } = useMemo(
-    () =>
-      computeFinykSchedule({
-        subscriptions,
-        manualDebts,
-        receivables,
-        transactions,
-        todayStart,
-      }),
-    [subscriptions, manualDebts, receivables, transactions, todayStart],
-  );
-
-  // Quick-action helpers: open the relevant section + reveal its form in a
-  // single tap, so the user doesn't expand → scroll → tap "+ Додати".
-  const openSubscriptionForm = () => {
-    setOpen((v) => ({ ...v, subscriptions: true }));
-    setShowSubForm(true);
-  };
-  const openAssetForm = () => {
-    setOpen((v) => ({ ...v, assets: true }));
-    setShowAssetForm(true);
-  };
-  const openDebtForm = () => {
-    setOpen((v) => ({ ...v, liabilities: true }));
-    setShowDebtForm(true);
-  };
-
-  useEffect(() => {
-    if (!showAssetForm || !open.assets) return;
-    const frame = requestAnimationFrame(() => {
-      assetFormRef.current?.scrollIntoView({
-        behavior: "smooth",
-        block: "start",
-      });
-      try {
-        assetNameInputRef.current?.focus({ preventScroll: true });
-      } catch {
-        assetNameInputRef.current?.focus();
-      }
-    });
-    return () => cancelAnimationFrame(frame);
-  }, [showAssetForm, open.assets]);
-
-  useEffect(() => {
-    if (!showDebtForm || !open.liabilities) return;
-    const frame = requestAnimationFrame(() => {
-      debtFormRef.current?.scrollIntoView({
-        behavior: "smooth",
-        block: "start",
-      });
-      try {
-        debtNameInputRef.current?.focus({ preventScroll: true });
-      } catch {
-        debtNameInputRef.current?.focus();
-      }
-    });
-    return () => cancelAnimationFrame(frame);
-  }, [showDebtForm, open.liabilities]);
-
-  if (txPicker) {
+  if (state.txPicker) {
     return (
       <AssetsTxPickerView
-        txPicker={txPicker}
-        setTxPicker={setTxPicker}
-        accounts={accounts}
-        transactions={transactions}
-        monoDebtLinkedTxIds={monoDebtLinkedTxIds}
-        toggleMonoDebtTx={toggleMonoDebtTx}
-        subscriptions={subscriptions}
-        updateSubscription={updateSubscription}
-        manualDebts={manualDebts}
-        receivables={receivables}
-        toggleLinkedTx={toggleLinkedTx}
-        showBalance={showBalance}
-        customCategories={customCategories}
+        txPicker={state.txPicker}
+        setTxPicker={state.setTxPicker}
+        accounts={state.accounts}
+        transactions={state.transactions}
+        monoDebtLinkedTxIds={state.monoDebtLinkedTxIds}
+        toggleMonoDebtTx={state.toggleMonoDebtTx}
+        subscriptions={state.subscriptions}
+        updateSubscription={state.updateSubscription}
+        manualDebts={state.manualDebts}
+        receivables={state.receivables}
+        toggleLinkedTx={state.toggleLinkedTx}
+        showBalance={state.showBalance}
+        customCategories={state.customCategories}
       />
     );
   }
@@ -227,645 +33,7 @@ export function Assets({
   return (
     <div className="flex-1 overflow-y-auto">
       <div className="max-w-4xl mx-auto px-4 pt-4 page-tabbar-pad space-y-1">
-        {/* Networth */}
-        <div className="bg-gradient-to-br from-emerald-600 via-emerald-700 to-emerald-900 text-white rounded-2xl p-5 mb-3 border border-white/10 shadow-float">
-          <div className="text-xs text-emerald-100/90 mb-1">
-            Загальний нетворс
-          </div>
-          <div
-            className={cn(
-              "text-3xl font-extrabold tracking-tight",
-              !showBalance && "tracking-widest",
-            )}
-          >
-            {showBalance
-              ? `${networth.toLocaleString("uk-UA", { maximumFractionDigits: 0 })} ₴`
-              : "••••••"}
-          </div>
-          <div className="text-xs text-emerald-100/85 mt-1">
-            {showBalance ? (
-              <>
-                Активи:{" "}
-                {totalAssets.toLocaleString("uk-UA", {
-                  maximumFractionDigits: 0,
-                })}{" "}
-                ₴ · Пасиви: −
-                {totalDebt.toLocaleString("uk-UA", {
-                  maximumFractionDigits: 0,
-                })}{" "}
-                ₴
-              </>
-            ) : (
-              "Суми приховано"
-            )}
-          </div>
-          {/* D — proportional assets/liabilities bar. Only rendered when the
-              user has data in both buckets, so a fresh account isn't greeted
-              with an empty sliver. */}
-          {showBalance && totalAssets + totalDebt > 0 && (
-            <AssetsLiabilitiesBar
-              assets={totalAssets}
-              liabilities={totalDebt}
-            />
-          )}
-        </div>
-
-        {/* C — compact stats strip. Surfaces numbers that would otherwise be
-            hidden inside collapsed sections: subs monthly total, next due
-            charge, biggest liability with a deadline. Horizontal scroll on
-            mobile, grid on wider screens. Only tiles with data render, so
-            the strip disappears entirely for empty accounts. */}
-        {/* На сторінці Активи залишаємо тільки «Пасив з дедлайном» —
-            «Сума підписок» + «Наступний платіж» переїхали на Планування
-            (PR #560), тут дублювати не треба. */}
-        <FinykStatsStrip
-          subsMonthly={0}
-          subsCount={0}
-          nextCharge={null}
-          urgentLiability={urgentLiability}
-          todayStart={todayStart}
-          showBalance={showBalance}
-          onOpenLiabilities={() =>
-            setOpen((v) => ({ ...v, liabilities: true }))
-          }
-          className="mb-3"
-        />
-
-        {/* E — quick-action CTAs. Each opens the respective section and its
-            inline form in one tap, collapsing the "expand → scroll → tap +"
-            flow into a single gesture. */}
-        <div className="grid grid-cols-3 gap-2 mb-3">
-          <QuickActionButton
-            iconName="refresh-cw"
-            label="Підписка"
-            onClick={openSubscriptionForm}
-          />
-          <QuickActionButton
-            iconName="trending-up"
-            label="Актив"
-            onClick={openAssetForm}
-          />
-          <QuickActionButton
-            iconName="trending-down"
-            label="Пасив"
-            onClick={openDebtForm}
-          />
-        </div>
-
-        {/* Auto-detected recurring suggestions (renders nothing if empty) */}
-        <RecurringSuggestions
-          transactions={transactions}
-          subscriptions={subscriptions}
-          dismissedRecurring={dismissedRecurring}
-          excludedTxIds={excludedTxIds}
-          onAdd={(candidate) => addSubscriptionFromRecurring?.(candidate)}
-          onDismiss={(key) => dismissRecurring?.(key)}
-        />
-
-        {/* Subscriptions section */}
-        <SectionBar
-          title="Підписки"
-          iconName="refresh-cw"
-          summary={`${subscriptions.length} активн${
-            subscriptions.length === 1 ? "а" : "их"
-          }`}
-          open={open.subscriptions}
-          onToggle={() =>
-            setOpen((v) => ({ ...v, subscriptions: !v.subscriptions }))
-          }
-        />
-        {open.subscriptions && (
-          <div className="mb-3 space-y-0">
-            {subscriptions.length > 0 && (
-              <button
-                type="button"
-                onClick={() => openHubModule("routine", "")}
-                className="w-full text-xs text-muted hover:text-text transition-colors pb-2 flex items-center justify-center gap-1.5"
-              >
-                <Icon name="calendar" size={14} aria-hidden />
-                <span>Побачити у календарі Рутини</span>
-                <Icon name="chevron-right" size={14} aria-hidden />
-              </button>
-            )}
-            {subscriptions.map((sub, i) => (
-              <SubCard
-                key={sub.id}
-                sub={sub}
-                transactions={transactions}
-                onDelete={() => {
-                  setSubscriptions((ss) => ss.filter((_, j) => j !== i));
-                  notifyFinykRoutineCalendarSync();
-                }}
-                onEdit={(updated) => {
-                  setSubscriptions((ss) =>
-                    ss.map((s, j) => (j === i ? { ...s, ...updated } : s)),
-                  );
-                  notifyFinykRoutineCalendarSync();
-                }}
-                onLinkTransactions={() =>
-                  setTxPicker({ type: "sub", subId: sub.id })
-                }
-              />
-            ))}
-            {showSubForm ? (
-              <Card variant="flat" radius="md" className="space-y-3 mt-2">
-                <Input
-                  placeholder="Назва"
-                  value={newSub.name}
-                  onChange={(e) =>
-                    setNewSub((a) => ({ ...a, name: e.target.value }))
-                  }
-                />
-                <Input
-                  placeholder="Ключове слово з транзакції"
-                  value={newSub.keyword}
-                  onChange={(e) =>
-                    setNewSub((a) => ({ ...a, keyword: e.target.value }))
-                  }
-                />
-                <Input
-                  placeholder="День списання (1-31)"
-                  type="number"
-                  min="1"
-                  max="31"
-                  value={newSub.billingDay}
-                  onChange={(e) =>
-                    setNewSub((a) => ({
-                      ...a,
-                      billingDay: Number(e.target.value),
-                    }))
-                  }
-                />
-                <div className="flex gap-2">
-                  <Button
-                    className="flex-1"
-                    size="sm"
-                    onClick={() => {
-                      if (newSub.name && newSub.billingDay) {
-                        setSubscriptions((ss) => [
-                          ...ss,
-                          { ...newSub, id: Date.now().toString() },
-                        ]);
-                        notifyFinykRoutineCalendarSync();
-                        setNewSub({
-                          name: "",
-                          emoji: "📱",
-                          keyword: "",
-                          billingDay: "",
-                          currency: "UAH",
-                        });
-                        setShowSubForm(false);
-                      }
-                    }}
-                  >
-                    Додати
-                  </Button>
-                  <Button
-                    className="flex-1"
-                    size="sm"
-                    variant="ghost"
-                    onClick={() => setShowSubForm(false)}
-                  >
-                    Скасувати
-                  </Button>
-                </div>
-              </Card>
-            ) : (
-              <button
-                onClick={() => setShowSubForm(true)}
-                className="w-full py-2.5 text-sm text-muted border border-dashed border-line rounded-xl hover:border-primary hover:text-primary transition-colors mt-2"
-              >
-                + Додати підписку
-              </button>
-            )}
-          </div>
-        )}
-
-        {/* Assets section */}
-        <SectionBar
-          title="Активи"
-          iconName="trending-up"
-          iconTone="success"
-          summary={`+${totalAssets.toLocaleString("uk-UA", {
-            maximumFractionDigits: 0,
-          })} ₴`}
-          open={open.assets}
-          onToggle={() => setOpen((v) => ({ ...v, assets: !v.assets }))}
-        />
-        {open.assets && (
-          <div className="mb-3 space-y-2">
-            <SectionHeading as="div" size="sm" className="pt-1">
-              <span className="inline-flex items-center gap-1.5">
-                <Icon name="credit-card" size={14} className="text-muted" />
-                Картки Monobank
-              </span>
-            </SectionHeading>
-            {accounts
-              .filter((a) => !hiddenAccounts.includes(a.id))
-              .map((a, i) => (
-                <div
-                  key={i}
-                  className="flex items-center justify-between py-2.5 px-1 border-b border-line last:border-0"
-                >
-                  <div className="flex items-center gap-3">
-                    <span
-                      className="inline-flex h-8 w-8 items-center justify-center rounded-xl bg-surface-muted text-muted"
-                      aria-hidden
-                    >
-                      <Icon name="credit-card" size={16} />
-                    </span>
-                    <div>
-                      <div className="text-sm font-medium">
-                        {getAccountLabel(a)}
-                      </div>
-                      <div className="text-xs text-subtle mt-0.5">
-                        {(a.balance / 100).toLocaleString("uk-UA", {
-                          minimumFractionDigits: 2,
-                        })}{" "}
-                        {a.currencyCode === 980
-                          ? "₴"
-                          : a.currencyCode === 840
-                            ? "$"
-                            : "€"}
-                      </div>
-                    </div>
-                  </div>
-                </div>
-              ))}
-
-            <SectionHeading as="div" size="sm" className="pt-2">
-              <span className="inline-flex items-center gap-1.5">
-                <Icon name="hand-coins" size={14} className="text-success" />
-                Мені винні
-              </span>
-            </SectionHeading>
-            {receivables.map((r) => (
-              <DebtCard
-                key={r.id}
-                name={r.name}
-                emoji={r.emoji}
-                remaining={calcReceivableRemaining(r, transactions)}
-                paid={getRecvPaid(r, transactions)}
-                total={getReceivableEffectiveTotal(r, transactions)}
-                dueDate={r.dueDate}
-                isReceivable
-                onDelete={() =>
-                  setReceivables((rs) => rs.filter((x) => x.id !== r.id))
-                }
-                onLink={() => setTxPicker({ id: r.id, type: "recv" })}
-                linkedCount={r.linkedTxIds?.length || 0}
-              />
-            ))}
-            {showRecvForm ? (
-              <Card variant="flat" radius="md" className="space-y-3">
-                <Input
-                  placeholder="Ім'я або назва"
-                  value={newRecv.name}
-                  onChange={(e) =>
-                    setNewRecv((a) => ({ ...a, name: e.target.value }))
-                  }
-                />
-                <Input
-                  placeholder="Сума ₴"
-                  type="number"
-                  value={newRecv.amount}
-                  onChange={(e) =>
-                    setNewRecv((a) => ({ ...a, amount: e.target.value }))
-                  }
-                />
-                <Input
-                  placeholder="Нотатка (необов'язково)"
-                  value={newRecv.note}
-                  onChange={(e) =>
-                    setNewRecv((a) => ({ ...a, note: e.target.value }))
-                  }
-                />
-                <Input
-                  type="date"
-                  value={newRecv.dueDate}
-                  onChange={(e) =>
-                    setNewRecv((a) => ({ ...a, dueDate: e.target.value }))
-                  }
-                />
-                <div className="flex gap-2">
-                  <Button
-                    className="flex-1"
-                    size="sm"
-                    onClick={() => {
-                      if (newRecv.name && newRecv.amount) {
-                        setReceivables((rs) => [
-                          ...rs,
-                          {
-                            ...newRecv,
-                            id: Date.now().toString(),
-                            amount: Number(newRecv.amount),
-                            linkedTxIds: [],
-                          },
-                        ]);
-                        setNewRecv({
-                          name: "",
-                          emoji: "👤",
-                          amount: "",
-                          note: "",
-                          dueDate: "",
-                        });
-                        setShowRecvForm(false);
-                      }
-                    }}
-                  >
-                    Додати
-                  </Button>
-                  <Button
-                    className="flex-1"
-                    size="sm"
-                    variant="ghost"
-                    onClick={() => setShowRecvForm(false)}
-                  >
-                    Скасувати
-                  </Button>
-                </div>
-              </Card>
-            ) : (
-              <button
-                onClick={() => setShowRecvForm(true)}
-                className="w-full py-2.5 text-sm text-muted border border-dashed border-line rounded-xl hover:border-primary hover:text-primary transition-colors"
-              >
-                + Додати актив «мені винні»
-              </button>
-            )}
-
-            <SectionHeading as="div" size="sm" className="pt-2">
-              <span className="inline-flex items-center gap-1.5">
-                <Icon name="piggy-bank" size={14} className="text-muted" />
-                Інші активи
-              </span>
-            </SectionHeading>
-            {showAssetForm ? (
-              <Card
-                ref={assetFormRef}
-                variant="finyk-soft"
-                radius="md"
-                className="space-y-3"
-              >
-                <div>
-                  <div className="text-sm font-bold text-text">Новий актив</div>
-                  <div className="text-xs text-muted mt-0.5">
-                    Готівка, брокерський рахунок, крипта тощо.
-                  </div>
-                </div>
-                <Input
-                  ref={assetNameInputRef}
-                  placeholder="Назва"
-                  value={newAsset.name}
-                  onChange={(e) =>
-                    setNewAsset((a) => ({ ...a, name: e.target.value }))
-                  }
-                />
-                <Input
-                  placeholder="Сума"
-                  type="number"
-                  value={newAsset.amount}
-                  onChange={(e) =>
-                    setNewAsset((a) => ({ ...a, amount: e.target.value }))
-                  }
-                />
-                <select
-                  className="input-focus-finyk w-full h-11 rounded-2xl border border-line bg-panelHi px-4 text-text"
-                  value={newAsset.currency}
-                  onChange={(e) =>
-                    setNewAsset((a) => ({ ...a, currency: e.target.value }))
-                  }
-                >
-                  <option>UAH</option>
-                  <option>USD</option>
-                  <option>EUR</option>
-                  <option>BTC</option>
-                </select>
-                <div className="flex gap-2">
-                  <Button
-                    className="flex-1"
-                    size="sm"
-                    onClick={() => {
-                      if (newAsset.name && newAsset.amount) {
-                        setManualAssets((a) => [...a, newAsset]);
-                        setNewAsset({
-                          name: "",
-                          amount: "",
-                          currency: "UAH",
-                          emoji: "💰",
-                        });
-                        setShowAssetForm(false);
-                      }
-                    }}
-                  >
-                    Додати
-                  </Button>
-                  <Button
-                    className="flex-1"
-                    size="sm"
-                    variant="ghost"
-                    onClick={() => setShowAssetForm(false)}
-                  >
-                    Скасувати
-                  </Button>
-                </div>
-              </Card>
-            ) : (
-              <button
-                onClick={() => setShowAssetForm(true)}
-                className="w-full py-2.5 text-sm text-muted border border-dashed border-line rounded-xl hover:border-primary hover:text-primary transition-colors"
-              >
-                + Додати актив
-              </button>
-            )}
-            {manualAssets.map((a, i) => (
-              <div
-                key={i}
-                className="flex items-center justify-between py-2.5 border-b border-text"
-              >
-                <div className="flex items-center gap-3">
-                  <span className="text-xl leading-none">{a.emoji}</span>
-                  <div>
-                    <div className="text-sm font-medium">{a.name}</div>
-                    <div className="text-xs text-subtle">{a.currency}</div>
-                  </div>
-                </div>
-                <div className="flex items-center gap-2">
-                  <span className="text-sm font-bold text-success">
-                    {Number(a.amount).toLocaleString("uk-UA")}{" "}
-                    {a.currency === "UAH"
-                      ? "₴"
-                      : a.currency === "USD"
-                        ? "$"
-                        : a.currency}
-                  </span>
-                  <button
-                    onClick={() =>
-                      setManualAssets((as) => as.filter((_, j) => j !== i))
-                    }
-                    className="text-subtle hover:text-danger text-sm transition-colors"
-                  >
-                    🗑
-                  </button>
-                </div>
-              </div>
-            ))}
-          </div>
-        )}
-
-        {/* Liabilities section */}
-        <SectionBar
-          title="Пасиви"
-          iconName="trending-down"
-          iconTone="danger"
-          summary={`−${totalDebt.toLocaleString("uk-UA", {
-            maximumFractionDigits: 0,
-          })} ₴`}
-          open={open.liabilities}
-          onToggle={() =>
-            setOpen((v) => ({ ...v, liabilities: !v.liabilities }))
-          }
-        />
-        {open.liabilities && (
-          <div className="mb-3 space-y-0">
-            {showDebtForm ? (
-              <Card
-                ref={debtFormRef}
-                variant="finyk-soft"
-                radius="md"
-                className="space-y-3 mb-2"
-              >
-                <div>
-                  <div className="text-sm font-bold text-text">Новий пасив</div>
-                  <div className="text-xs text-muted mt-0.5">
-                    Кредит, борг або інше зобовʼязання.
-                  </div>
-                </div>
-                <div className="flex gap-2">
-                  <Input
-                    ref={debtNameInputRef}
-                    className="flex-1"
-                    placeholder="Назва пасиву (кредит, борг…)"
-                    value={newDebt.name}
-                    onChange={(e) =>
-                      setNewDebt((a) => ({ ...a, name: e.target.value }))
-                    }
-                  />
-                  <VoiceMicButton
-                    size="md"
-                    label="Голосовий ввід"
-                    onResult={(transcript) => {
-                      const parsed = parseExpenseVoice(transcript);
-                      if (!parsed) return;
-                      setNewDebt((a) => ({
-                        ...a,
-                        name: parsed.name || a.name,
-                        totalAmount:
-                          parsed.amount != null
-                            ? String(Math.round(parsed.amount))
-                            : a.totalAmount,
-                      }));
-                    }}
-                  />
-                </div>
-                <Input
-                  placeholder="Загальна сума ₴"
-                  type="number"
-                  value={newDebt.totalAmount}
-                  onChange={(e) =>
-                    setNewDebt((a) => ({ ...a, totalAmount: e.target.value }))
-                  }
-                />
-                <Input
-                  type="date"
-                  value={newDebt.dueDate}
-                  onChange={(e) =>
-                    setNewDebt((a) => ({ ...a, dueDate: e.target.value }))
-                  }
-                />
-                <div className="flex gap-2">
-                  <Button
-                    className="flex-1"
-                    size="sm"
-                    onClick={() => {
-                      if (newDebt.name && newDebt.totalAmount) {
-                        setManualDebts((ds) => [
-                          ...ds,
-                          {
-                            ...newDebt,
-                            id: Date.now().toString(),
-                            totalAmount: Number(newDebt.totalAmount),
-                            linkedTxIds: [],
-                          },
-                        ]);
-                        setNewDebt({
-                          name: "",
-                          emoji: "💸",
-                          totalAmount: "",
-                          dueDate: "",
-                        });
-                        setShowDebtForm(false);
-                      }
-                    }}
-                  >
-                    Додати
-                  </Button>
-                  <Button
-                    className="flex-1"
-                    size="sm"
-                    variant="ghost"
-                    onClick={() => setShowDebtForm(false)}
-                  >
-                    Скасувати
-                  </Button>
-                </div>
-              </Card>
-            ) : (
-              <button
-                onClick={() => setShowDebtForm(true)}
-                className="w-full py-2.5 text-sm text-muted border border-dashed border-line rounded-xl hover:border-primary hover:text-primary transition-colors mb-2"
-              >
-                + Додати пасив
-              </button>
-            )}
-            {monoDebtAccounts.map((a, i) => {
-              const linkedIds = monoDebtLinkedTxIds[a.id] || [];
-              const paidFromLinked = transactions
-                .filter((t) => linkedIds.includes(t.id))
-                .reduce((s, t) => s + Math.abs(t.amount / 100), 0);
-              const remaining = getMonoDebt(a);
-              const volatileTotal = paidFromLinked + remaining;
-              return (
-                <DebtCard
-                  key={i}
-                  name={getAccountLabel(a)}
-                  emoji="🖤"
-                  remaining={remaining}
-                  paid={paidFromLinked}
-                  total={volatileTotal}
-                  onLink={() => setTxPicker({ id: a.id, type: "monoDebt" })}
-                  linkedCount={linkedIds.length}
-                />
-              );
-            })}
-            {manualDebts.map((d) => (
-              <DebtCard
-                key={d.id}
-                name={d.name}
-                emoji={d.emoji}
-                remaining={calcDebtRemaining(d, transactions)}
-                paid={getDebtPaid(d, transactions)}
-                total={getDebtEffectiveTotal(d, transactions)}
-                dueDate={d.dueDate}
-                onDelete={() =>
-                  setManualDebts((ds) => ds.filter((x) => x.id !== d.id))
-                }
-                onLink={() => setTxPicker({ id: d.id, type: "debt" })}
-                linkedCount={d.linkedTxIds?.length || 0}
-              />
-            ))}
-          </div>
-        )}
+        <AssetsTable state={state} />
       </div>
     </div>
   );
